Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
google-generativeai==0.3.2
groq==0.4.1
python-dotenv==1.0.0
requests==2.31.0
python-jose[cryptography]==3.3.0
fastapi-cors
# Added to fix dependency conflict
httpx==0.24.1
httpcore==0.17.3